+++ hutool-core/src/main/java/cn/hutool/core/lang/Snowflake.java
@@ -1,6 +1,7 @@
 package cn.hutool.core.lang;
 
 import cn.hutool.core.date.SystemClock;
+import cn.hutool.core.util.StrUtil;
 
 /**
  * Twitter的Snowflake 算法<br>
@@ -25,25 +26,33 @@ import cn.hutool.core.date.SystemClock;
  */
 public class Snowflake {
 
+	//Thu, 04 Nov 2010 01:42:54 GMT
 	private final long twepoch = 1288834974657L;
 	private final long workerIdBits = 5L;
 	private final long datacenterIdBits = 5L;
+	////最大支持机器节点数0~31，一共32个
 	private final long maxWorkerId = -1L ^ (-1L << workerIdBits);
+	//最大支持数据中心节点数0~31，一共32个
 	private final long maxDatacenterId = -1L ^ (-1L << datacenterIdBits);
+	//序列号12位
 	private final long sequenceBits = 12L;
+	//机器节点左移12位
 	private final long workerIdShift = sequenceBits;
+	//数据中心节点左移17位
 	private final long datacenterIdShift = sequenceBits + workerIdBits;
+	//时间毫秒数左移22位
 	private final long timestampLeftShift = sequenceBits + workerIdBits + datacenterIdBits;
-	private final long sequenceMask = -1L ^ (-1L << sequenceBits);
+	private final long sequenceMask = -1L ^ (-1L << sequenceBits);//4095
 
 	private long workerId;
 	private long datacenterId;
 	private long sequence = 0L;
 	private long lastTimestamp = -1L;
 	private boolean useSystemClock;
-	
+
 	/**
 	 * 构造
+	 * 
 	 * @param workerId 终端ID
 	 * @param datacenterId 数据中心ID
 	 */
@@ -53,16 +62,17 @@ public class Snowflake {
 
 	/**
 	 * 构造
+	 * 
 	 * @param workerId 终端ID
 	 * @param datacenterId 数据中心ID
 	 * @param isUseSystemClock 是否使用{@link SystemClock} 获取当前时间戳
 	 */
 	public Snowflake(long workerId, long datacenterId, boolean isUseSystemClock) {
 		if (workerId > maxWorkerId || workerId < 0) {
-			throw new IllegalArgumentException(String.format("worker Id can't be greater than %d or less than 0", maxWorkerId));
+			throw new IllegalArgumentException(StrUtil.format("worker Id can't be greater than {} or less than 0", maxWorkerId));
 		}
 		if (datacenterId > maxDatacenterId || datacenterId < 0) {
-			throw new IllegalArgumentException(String.format("datacenter Id can't be greater than %d or less than 0", maxDatacenterId));
+			throw new IllegalArgumentException(StrUtil.format("datacenter Id can't be greater than {} or less than 0", maxDatacenterId));
 		}
 		this.workerId = workerId;
 		this.datacenterId = datacenterId;
@@ -71,12 +81,14 @@ public class Snowflake {
 
 	/**
 	 * 下一个ID
+	 * 
 	 * @return ID
 	 */
 	public synchronized long nextId() {
-		long timestamp = useSystemClock ? SystemClock.now() : System.currentTimeMillis();
+		long timestamp = genTime();
 		if (timestamp < lastTimestamp) {
-			throw new RuntimeException(String.format("Clock moved backwards.  Refusing to generate id for %d milliseconds", lastTimestamp - timestamp));
+			 //如果服务器时间有问题(时钟后退) 报错。
+			throw new IllegalStateException(StrUtil.format("Clock moved backwards. Refusing to generate id for {}ms", lastTimestamp - timestamp));
 		}
 		if (lastTimestamp == timestamp) {
 			sequence = (sequence + 1) & sequenceMask;
@@ -92,11 +104,27 @@ public class Snowflake {
 		return ((timestamp - twepoch) << timestampLeftShift) | (datacenterId << datacenterIdShift) | (workerId << workerIdShift) | sequence;
 	}
 
+	//------------------------------------------------------------------------------------------------------------------------------------ Private method start
+	/**
+	 * 循环等待下一个时间
+	 * 
+	 * @param lastTimestamp 上次记录的时间
+	 * @return 下一个时间
+	 */
 	private long tilNextMillis(long lastTimestamp) {
-		long timestamp = useSystemClock ? SystemClock.now() : System.currentTimeMillis();
+		long timestamp = genTime();
 		while (timestamp <= lastTimestamp) {
-			timestamp = useSystemClock ? SystemClock.now() : System.currentTimeMillis();
+			timestamp = genTime();
 		}
 		return timestamp;
 	}
+	
+	/**
+	 * 生成时间戳
+	 * @return 时间戳
+	 */
+	private long genTime() {
+		return this.useSystemClock ? SystemClock.now() : System.currentTimeMillis();
+	}
+	//------------------------------------------------------------------------------------------------------------------------------------ Private method end
 }