+++ hutool-core/src/test/java/cn/hutool/core/text/UnicodeUtilTest.java
@@ -39,4 +39,11 @@ public class UnicodeUtilTest {
 		String res = UnicodeUtil.toString(str);
 		Assert.assertEquals("aaa中文\\u111\\urtyu&", res);
 	}
+	
+	@Test
+	public void convertTest5() {
+		String str = "{\"code\":403,\"enmsg\":\"Product not found\",\"cnmsg\":\"\\u4ea7\\u54c1\\u4e0d\\u5b58\\u5728\\uff0c\\u6216\\u5df2\\u5220\\u9664\",\"data\":null}";
+		String res = UnicodeUtil.toString(str);
+		Assert.assertEquals("{\"code\":403,\"enmsg\":\"Product not found\",\"cnmsg\":\"产品不存在，或已删除\",\"data\":null}", res);
+	}
 }