+++ hutool-crypto/src/main/java/cn/hutool/crypto/symmetric/SymmetricCrypto.java
@@ -17,6 +17,7 @@ import cn.hutool.core.util.CharsetUtil;
 import cn.hutool.core.util.HexUtil;
 import cn.hutool.core.util.RandomUtil;
 import cn.hutool.core.util.StrUtil;
+import cn.hutool.crypto.BouncyCastleSupport;
 import cn.hutool.crypto.CryptoException;
 import cn.hutool.crypto.SecureUtil;
 
@@ -29,7 +30,7 @@ import cn.hutool.crypto.SecureUtil;
  * @author Looly
  *
  */
-public class SymmetricCrypto {
+public class SymmetricCrypto extends BouncyCastleSupport{
 
 	/** SecretKey 负责保存对称密钥 */
 	private SecretKey secretKey;