+++ hutool-http/src/test/java/cn/hutool/http/test/HttpUtilTest.java
@@ -20,7 +20,7 @@ public class HttpUtilTest {
 	@Test
 	@Ignore
 	public void postTest() {
-		String result = HttpUtil.createPost("http://api.uhaozu.com/goods/description/1120448506").charset(CharsetUtil.UTF_8).execute().body();
+		String result = HttpUtil.createPost("api.uhaozu.com/goods/description/1120448506").charset(CharsetUtil.UTF_8).execute().body();
 		Console.log(result);
 	}