+++ hutool-log/src/main/java/cn/hutool/log/dialect/jboss/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * jboss-logging实现
+ * 
+ * @author looly
+ *
+ */
+package cn.hutool.log.dialect.jboss;